backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentlake"
version = "0.1.0"
description = "Distributed, agent-friendly data lake"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "minio>=7.2",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "PyMuPDF>=1.24",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "python-pptx>=0.6",
    "tiktoken>=0.7",
    "pyyaml>=6.0",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0",
    "websockets>=12.0",
    "anthropic>=0.30",
    "langgraph>=1.0",
    "wsgidav>=4.3",
    "cheroot>=11.0",
    "langchain-core>=0.3",
    "pgvector>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "respx>=0.21",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentlake"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: tests requiring running infrastructure (PostgreSQL, Redis, MinIO)",
    "scale: performance and load tests",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["agentlake"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 85
